Fixed build error and setup structured data for SEO
