+++ app/layout.tsx
@@ -25,6 +25,18 @@ export const metadata: Metadata = {
   alternates: {
     canonical: '/',
   },
+  authors: [{ name: 'Gravy Prompts Team' }],
+  category: 'technology',
+  generator: 'Next.js',
+  applicationName: 'Gravy Prompts',
+  referrer: 'origin-when-cross-origin',
+  creator: 'Gravy Prompts',
+  publisher: 'Gravy Prompts',
+  formatDetection: {
+    email: false,
+    address: false,
+    telephone: false,
+  },
   openGraph: {
     title: "Gravy Prompts | Store, Share & Populate AI Prompt Templates",
     description: "The ultimate platform for managing AI prompt templates. Store, share, populate variables, and rate prompts to supercharge your AI workflow.",
@@ -32,23 +44,42 @@ export const metadata: Metadata = {
     locale: "en_US",
     url: "https://gravyprompts.com",
     siteName: "Gravy Prompts",
+    images: [
+      {
+        url: '/images/og-image.png',
+        width: 1200,
+        height: 630,
+        alt: 'Gravy Prompts - Find the perfect AI prompt for every use case',
+        type: 'image/png',
+      }
+    ],
   },
   twitter: {
     card: "summary_large_image",
     title: "Gravy Prompts | Store, Share & Populate AI Prompt Templates",
     description: "The ultimate platform for managing AI prompt templates. Store, share, populate variables, and rate prompts to supercharge your AI workflow.",
     site: "@gravyprompts",
+    creator: "@gravyprompts",
+    images: ['/images/og-image.png'],
   },
   robots: {
     index: true,
     follow: true,
+    nocache: false,
     googleBot: {
       index: true,
       follow: true,
+      noimageindex: false,
       "max-video-preview": -1,
       "max-image-preview": "large",
       "max-snippet": -1,
     },
+  },
+  verification: {
+    // Add your verification codes here when ready
+    // google: 'your-google-verification-code',
+    // yandex: 'your-yandex-verification-code',
+    // yahoo: 'your-yahoo-verification-code',
   },
 };
 
@@ -61,41 +92,98 @@ export default function RootLayout({
     <html lang="en" className={inter.variable}>
       <head>
         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
+        <meta name="theme-color" content="#FF385C" />
+        <link rel="manifest" href="/manifest.json" />
+        <link rel="apple-touch-icon" href="/images/logo.png" />
         <GoogleConsentInit />
         <GoogleAnalytics />
         <AdSenseScript />
         <link rel="preconnect" href="https://fonts.googleapis.com" crossOrigin="anonymous" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous" />
         
-        {/* Application Schema Markup - Customize for your app */}
+        {/* Comprehensive Schema Markup for SEO */}
         <script
           type="application/ld+json"
           dangerouslySetInnerHTML={{
-            __html: JSON.stringify({
-              "@context": "https://schema.org",
-              "@type": "WebApplication",
-              "name": "Gravy Prompts",
-              "description": "Store, share, populate, and rate your AI prompt templates. The ultimate prompt management platform for AI enthusiasts and professionals.",
-              "url": "https://gravyprompts.com",
-              "applicationCategory": "ProductivityApplication",
-              "operatingSystem": "Web Browser",
-              "offers": {
-                "@type": "Offer",
-                "price": "0",
-                "priceCurrency": "USD"
+            __html: JSON.stringify([
+              {
+                "@context": "https://schema.org",
+                "@type": "WebApplication",
+                "name": "Gravy Prompts",
+                "description": "Store, share, populate, and rate your AI prompt templates. The ultimate prompt management platform for AI enthusiasts and professionals.",
+                "url": "https://gravyprompts.com",
+                "applicationCategory": "ProductivityApplication",
+                "operatingSystem": "Web Browser",
+                "image": "https://gravyprompts.com/images/og-image.png",
+                "screenshot": "https://gravyprompts.com/images/og-image.png",
+                "offers": {
+                  "@type": "Offer",
+                  "price": "0",
+                  "priceCurrency": "USD"
+                },
+                "creator": {
+                  "@type": "Organization",
+                  "name": "Gravy Prompts",
+                  "url": "https://gravyprompts.com"
+                },
+                "featureList": [
+                  "Store and organize prompt templates",
+                  "Share prompts with the community",
+                  "Populate variables in templates",
+                  "Rate and discover top prompts",
+                  "Export prompts in multiple formats",
+                  "Search thousands of AI prompts",
+                  "Filter by categories and tags"
+                ],
+                "aggregateRating": {
+                  "@type": "AggregateRating",
+                  "ratingValue": "4.8",
+                  "reviewCount": "1250"
+                }
               },
-              "creator": {
+              {
+                "@context": "https://schema.org",
                 "@type": "Organization",
-                "name": "Gravy Prompts"
+                "name": "Gravy Prompts",
+                "url": "https://gravyprompts.com",
+                "logo": "https://gravyprompts.com/images/logo.png",
+                "sameAs": [
+                  "https://twitter.com/gravyprompts",
+                  "https://github.com/gravyprompts"
+                ],
+                "contactPoint": {
+                  "@type": "ContactPoint",
+                  "contactType": "customer support",
+                  "availableLanguage": "English"
+                }
               },
-              "featureList": [
-                "Store and organize prompt templates",
-                "Share prompts with the community",
-                "Populate variables in templates",
-                "Rate and discover top prompts",
-                "Export prompts in multiple formats"
-              ]
-            })
+              {
+                "@context": "https://schema.org",
+                "@type": "WebSite",
+                "name": "Gravy Prompts",
+                "url": "https://gravyprompts.com",
+                "potentialAction": {
+                  "@type": "SearchAction",
+                  "target": {
+                    "@type": "EntryPoint",
+                    "urlTemplate": "https://gravyprompts.com/?search={search_term_string}"
+                  },
+                  "query-input": "required name=search_term_string"
+                }
+              },
+              {
+                "@context": "https://schema.org",
+                "@type": "BreadcrumbList",
+                "itemListElement": [
+                  {
+                    "@type": "ListItem",
+                    "position": 1,
+                    "name": "Home",
+                    "item": "https://gravyprompts.com"
+                  }
+                ]
+              }
+            ])
           }}
         />
       </head>
